pyproject.toml
[tool.poetry]
name = "autorizer"
version = "0.1.0"
description = ""
authors = ["Tirik8 <yaroslavsemenoviop@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
pydantic = "^2.11.3"


[tool.poetry.group.dev.dependencies]
uvicorn = "0.22.0"
sqlalchemy = "2.0.15"
passlib = "1.7.4"
python-jose = "3.3.0"
python-multipart = "0.0.6"
python-dotenv = "1.0.0"
alembic = "1.13.1"
psycopg2-binary = "2.9.9"
pydantic = {extras = ["email"], version = "^2.11.3"}
pydantic-settings = "^2.9.1"
bcrypt = "^4.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
